refreshToken 만료 및 로그아웃 시 JWT 정리

+++ src/main/java/com/boot/sonorous/common/controller/api/SonorousLoginApiController.java
@@ -1,6 +1,7 @@
 package com.boot.sonorous.common.controller.api;
 import com.boot.sonorous.common.dto.SignInDto;
 import com.boot.sonorous.common.entity.JwtToken;
+import com.boot.sonorous.common.entity.RefreshToken;
 import com.boot.sonorous.common.service.SonorousMemberService;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
@@ -9,6 +10,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/api/auth")
@@ -68,26 +71,82 @@ public class SonorousLoginApiController {
             }
         }
         String newAccessToken="";
-        String newRefreshToken="";
         // refreshToken이 유효하면
         if(sonorousMemberService.isTokenExpired(refreshToken)){
-            sonorousMemberService.deleteRefreshToken(refreshToken);
             JwtToken jwtToken = sonorousMemberService.reissueToken(refreshToken);
             newAccessToken = jwtToken.getAccessToken();
-            newRefreshToken = jwtToken.getRefreshToken();
-
-            Cookie cookie = new Cookie("refreshToken", newRefreshToken);
-            cookie.setHttpOnly(true);
-            cookie.setSecure(true);
-            cookie.setPath("/");
-            cookie.setMaxAge(7 * 24 * 60 * 60);
-            response.addCookie(cookie);
-
-            sonorousMemberService.insertRefreshToken(newRefreshToken);
         }
-
         return ResponseEntity.ok(newAccessToken);
     }
 
+    @GetMapping("/refreshChk")
+    public ResponseEntity<?> checkRefreshToken(HttpServletRequest request, HttpServletResponse response) {
+        Cookie[] cookies = request.getCookies();
+        String refreshToken="";
+        String reply="";
+        if(cookies != null){
+            for (Cookie cookie : cookies){
+                if("refreshToken".equals(cookie.getName())){
+                    refreshToken = cookie.getValue();
+                }
+            }
+        }
+        if(refreshToken != null && !refreshToken.isEmpty()){
 
+            boolean result = sonorousMemberService.isTokenExpired(refreshToken);
+
+            if(!sonorousMemberService.isTokenExpired(refreshToken)){
+
+                //DB에서 만료된 refreshToken 삭제
+                sonorousMemberService.deleteRefreshToken(refreshToken);
+
+                // 쿠키에서 refreshToken 삭제
+                Cookie cookie = new Cookie("refreshToken", null);
+                cookie.setPath("/");
+                cookie.setMaxAge(0);
+                cookie.setHttpOnly(true);
+                cookie.setSecure(true);
+                response.addCookie(cookie);
+
+                reply = "Refresh token is expired";
+            }else{
+                reply = "Refresh token is valid";
+            }
+        }
+        return ResponseEntity.ok(reply);
+    }
+
+    @GetMapping("/logout")
+    public ResponseEntity<?> logout(HttpServletRequest request, HttpServletResponse response) {
+        Cookie[] cookies = request.getCookies();
+        String refreshToken="";
+        String reply="";
+        if(cookies != null){
+            for (Cookie cookie : cookies){
+                if("refreshToken".equals(cookie.getName())){
+                    refreshToken = cookie.getValue();
+                }
+            }
+        }
+
+        if(refreshToken != null && !refreshToken.isEmpty()) {
+            if (sonorousMemberService.isTokenExpired(refreshToken)) {
+                // DB에서 만료된 refreshToken 삭제
+                sonorousMemberService.deleteRefreshToken(refreshToken);
+
+                // 쿠키에서 refreshToken 삭제
+                Cookie cookie = new Cookie("refreshToken", null);
+                cookie.setPath("/");
+                cookie.setMaxAge(0);
+                cookie.setHttpOnly(true);
+                cookie.setSecure(true);
+                response.addCookie(cookie);
+
+                reply = "Logout completed";
+            } else {
+                reply = "Refresh token is expired";
+            }
+        }
+        return ResponseEntity.ok(reply);
+    }
 }
